Fix SettingsOption creating phantom Options
SettingsOption in its current form doesn't read in the closing tag, leaving a closing tag for the category to read, which the xmlReader doesn't recognize as a closing tag. Slightly changing SettingsOption to use ReadSubTree, lets it properly close the tag, rather the Option is a self closing tag or a normal tag with a separate closing tag.